Updated Front Page
I removed sections of text that mentioned the spring and removed all examples assuming that sites will have an established method now that the sprint is over.  If they need help they can refer back to the sprint github.  I left De-Ident alone knowing that it will change in the future.  I replaced the timeline and the goals.